rag-chatbot/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
beanie==1.25.0
weaviate-client==3.24.2
# openai==0.28.0
cohere==5.13.12
sentence-transformers==3.4.1
python-multipart==0.0.20
python-dotenv==1.0.1
# motor==3.3.2
# python-multipart==0.0.5
# httpx==0.23.0
pymupdf==1.25.2
# nltk==3.9.1
# regex==2024.11.6
# joblib==1.4.2
# Pillow==11.1.0
# numpy==2.2.2
# opencv-python==4.11.0.86


# Successfully installed SQLAlchemy-2.0.38 aiohappyeyeballs-2.4.6 aiohttp-3.11.12 aiosignal-1.3.2 attrs-25.1.0 frozenlist-1.5.0 greenlet-3.1.1 jsonpatch-1.33 jsonpointer-3.0.0 langchain-0.3.18 langchain-core-0.3.34 langchain-text-splitters-0.3.6 langsmith-0.3.8 multidict-6.1.0 orjson-3.10.15 pdfminer.six-20240706 propcache-0.2.1 pypdf-5.3.0 requests-toolbelt-1.0.0 tenacity-9.0.0 yarl-1.18.3 zstandard-0.23.0